Added mount demo. Cleaned up core
